+++ codex-rs/core/src/state/service.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::AuthManager;
 use crate::RolloutRecorder;
 use crate::mcp_connection_manager::McpConnectionManager;
-use crate::openai_models::models_manager::ModelsManager;
+use crate::models_manager::models_manager::ModelsManager;
 use crate::skills::SkillsManager;
 use crate::tools::sandboxing::ApprovalStore;
 use crate::unified_exec::UnifiedExecSessionManager;
