+++ codex-rs/core/src/tasks/mod.rs
@@ -19,7 +19,7 @@ use tracing::warn;
 use crate::AuthManager;
 use crate::codex::Session;
 use crate::codex::TurnContext;
-use crate::openai_models::models_manager::ModelsManager;
+use crate::models_manager::models_manager::ModelsManager;
 use crate::protocol::EventMsg;
 use crate::protocol::TaskCompleteEvent;
 use crate::protocol::TurnAbortReason;
