+++ codex-rs/core/tests/responses_headers.rs
@@ -10,7 +10,7 @@ use codex_core::Prompt;
 use codex_core::ResponseEvent;
 use codex_core::ResponseItem;
 use codex_core::WireApi;
-use codex_core::openai_models::models_manager::ModelsManager;
+use codex_core::models_manager::models_manager::ModelsManager;
 use codex_otel::otel_manager::OtelManager;
 use codex_protocol::ConversationId;
 use codex_protocol::config_types::ReasoningSummary;
