+++ codex-rs/tui2/src/chatwidget/tests.rs
@@ -10,7 +10,7 @@ use codex_core::CodexAuth;
 use codex_core::config::Config;
 use codex_core::config::ConfigBuilder;
 use codex_core::config::Constrained;
-use codex_core::openai_models::models_manager::ModelsManager;
+use codex_core::models_manager::models_manager::ModelsManager;
 use codex_core::protocol::AgentMessageDeltaEvent;
 use codex_core::protocol::AgentMessageEvent;
 use codex_core::protocol::AgentReasoningDeltaEvent;
